Changes how we add asdf to the PATH.
Rather than 'source'ing, we manually prepend the paths to PATH.

We do this because of how the '.bash_profile' is loaded when opening a VS Code terminal. For some reason, it runs it twice, in which case all of the other prepended paths will go in front of 'asdf', which casues problems.